Upgrade DXv20.1.2 prview and EF support added
